Highlighter is now format agnostic. MarkdownAutoFormat uses patterns from MarkdownHighlighterPattern.
